+++ CHANGELOG.md
@@ -18,6 +18,9 @@ release will be created for any release tested in production, and may be marked 
 
 ## Unreleased
 
+
+## [8.0.22] - 2021-11-15
+
 ### Transfer from Harvard @7.4.71.1 - 2021-11-15
 
 - [Added] column option for "choice_label" and ensure it works for all types of display and editing